==> 00_wistem_student_form.qmd <==
---
title: "WiSTEM 2022 - 2023"
subtitle: "Student Form Pre and Post Analyses"
format:
  html:
    toc: true
    embed-resources: true
    echo: false
    link-external-newwindow: true
execute:
  warning: false
  message: false
from: markdown+emoji  
editor: 
  markdown: 
    wrap: 72
editor_options: 
  chunk_output_type: console
---

```{r}
#| label: load-packages

library(tidyverse)
library(patchwork)
library(readr)
library(kableExtra)

set.seed(1968)
```

```{r}
#| label: load-data-sets

data <- read_csv(file = "data/processed/wistem_student_forms_clean.csv")
pre_data <- read_csv(file = "data/processed/wistem_pre_data.csv")
post_data <- read_csv(file = "data/processed/wistem_post_data.csv")

data$work_hard <- factor(data$work_hard, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))
data$knowledge <- factor(data$knowledge, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))
data$courses <- factor(data$courses, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))
data$succesful <- factor(data$succesful, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))
data$important <- factor(data$important, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))
data$member <- factor(data$member, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))
data$career <- factor(data$career, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))
data$identity <- factor(data$identity, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))
data$meaningful <- factor(data$meaningful, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))
data$everyday <- factor(data$everyday, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))
```

# Background

**WiSTEM:** In total, there were 17 overall WiSTEM meetings held throughout the 2022-2023 academic school year. However, there were 13 meetings with attendance data and 13 meetings with exit slip data. Some of these overlap in missing both attendance and exit slip data. 

**WiENG:** In total, there were 25 overall WiENG meetings held throughout the 2022-2023 academic school year. However, there were 21 meetings with attendance data while there were 18 meetings with exit slip data. Some of these overlap in missing both attendance and exit slip data. 

**Distribution:** As can be seen in the table below, there were a total of 74 student forms obtained for the Pre-Test while there were only 48 student forms obtained for the Post-Test. There is a large discrepancy of 26 less forms being filled out by the end of the year, but the general trend can still be identified. These forms were either filled out during the original WiSTEM meetings (where a majority of them were) while some were filled out during some WiENG meetings. 
```{r}
#| label: background

data %>%
  count(survey_type) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100
  ) %>%
  arrange(desc(n)) %>%
  rename(
    `Number of Forms` = n, 
    `Survey Type` = survey_type, 
    Proportion = proportion, 
    Percent = percent
  ) %>%
  kbl(caption = "Distribution of Pre and Post Forms") %>%
  kable_styling()
```

# Career Analysis
## Career Amount

**Question:** List up to ten (10) STE(A)M career or job possibilities. 

**Distribution:** As can be seen in the two distribution charts and single table below, there were a handful of students who did not answer the question while there were also a handful that wrote down 10 careers on the paper. In both cases, there is a greater proportion of student who did not fill out this question or wrote down 10 career options during the Post-Test at the end of the academic year. Either way, the mean for how many careers were written down was higher for the Pre-Test by 0.3 as the mean was 5.0 for the Pre-Test and 4.7 for the Post-Test. There isn't that big of a difference. The main difference is seen with the Pre-Test being a unimodal distribution with one peak at 5.0 while the Pre-Test is bimodal with a peak around 5.0 and another peak around 9.0, showing that there was a second majority of students putting down 9 careers.
```{r}
#| label: career-amount-compare

# Pre Test
pre_career_amount_plot <- ggplot(data = pre_data, mapping = aes(x = career_amount)) + 
  geom_density() + 
  geom_rug() + 
  theme_minimal() + 
  labs(
    x = "Number of Careers", 
    y = "Density", 
    title = "Pre Test", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023"
    ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

pre_career_amount_table <- pre_data %>%
  count(career_amount) %>%
  mutate(
    mean = mean(career_amount), 
    max = max(career_amount), 
    min = min(career_amount), 
    std_dev = sd(career_amount),
    median = median(career_amount),
    test = "pre"
  ) %>%
  select(test, min, median, mean, std_dev, max) %>%
  head(1)

# Post Test
post_career_amount_plot <- ggplot(data = post_data, mapping = aes(x = career_amount)) + 
  geom_density() + 
  geom_rug() + 
  theme_minimal() + 
  labs(
    x = "Number of Careers", 
    y = "Density", 
    title = "Post Test", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023"
    ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

post_career_amount_table <- post_data %>%
  count(career_amount) %>%
  mutate(
    mean = mean(career_amount), 
    max = max(career_amount), 
    min = min(career_amount), 
    std_dev = sd(career_amount),
    median = median(career_amount),
    test = "post"
  ) %>%
  select(test, min, median, mean, std_dev, max) %>%
  head(1)

# Compare 
## Plot
pre_career_amount_plot + post_career_amount_plot
```

::: {.callout-note icon=false collapse="true"}
## Tables for Career Amount Analysis
```{r}
# Table 1 (Statistical Analysis)
career_amount_table <- pre_career_amount_table %>%
  bind_rows(post_career_amount_table) %>%
  rename(
    Type = test, 
    Min = min, 
    Median = median, 
    Mean = mean, 
    `Standard Dev` = std_dev, 
    Max = max
  )

career_amount_table %>%
  kbl(caption = "Comparing Statistical Summaries of Career Amounts") %>%
  kable_styling()

# Table 2 (0 and 10 Career Amount Analysis)
## Pre Test
pre_0_10_career_amount <- pre_data %>%
  count(career_amount) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100
  ) %>%
  filter(career_amount == 0 | career_amount == 10) %>%
  mutate(test = "pre") %>%
  select(test, career_amount, proportion, percent)

## Post Test
post_0_10_career_amount <- post_data %>%
  count(career_amount) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100
  ) %>%
  filter(career_amount == 0 | career_amount == 10) %>%
  mutate(test = "post") %>%
  select(test, career_amount, proportion, percent)

## Compare
career_amount_0_10_table <- pre_0_10_career_amount %>%
  bind_rows(post_0_10_career_amount) %>%
  rename(
    Type = test, 
    `Number of Careers` = career_amount, 
    Proportion = proportion, 
    Percent = percent
  )

career_amount_0_10_table %>%
  arrange(`Number of Careers`) %>%
  kbl(caption = "Comparing 0 and 10 Career Amounts") %>%
  kable_styling()
```

:::

## Most and Least Popular Careers

**Distribution:** In the folded blue tabs below, the most and least commonly written majors were identified in the Pre and Post-Test responses. For the Pre-Test, the three most common were Engineer (8.04%), Chemist (6.97%), and Doctor (6.97%) while for the Post-Test, the three most common were Doctor (8.00%), Engineer (8.00%), but also Biologist (6.00%). In the third foldable tab, there are lists of unique career options that were only in the Pre-Test or only in the Post-Test as well as another list of careers that were present in both forms. 

::: {.callout-note icon=false collapse="true"}
## Pre Test: Most and Least Common
::: {.callout-tip icon=false collapse="true"}
## Most Common Careers in the Pre-Test
```{r}
#| label: pre-most-career-cat

# Background
pre_data_longer <- pre_data %>%
  select(survey_type, career_one, career_two, career_three, career_four, career_five, career_six, career_seven, career_eight, career_nine, career_ten) %>% 
  pivot_longer(!survey_type, names_to = "career_num", values_to = "career") %>%
  select(-career_num) %>%
  filter(!is.na(career))

pre_career <- pre_data_longer %>%
  count(career) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "pre"
  ) %>%
  select(test, career, n, proportion, percent) %>%
  rename(count = n)

# Most
pre_most_career <- pre_career %>%
  arrange(desc(count)) %>%
  head(10) %>%
  rename(
    Type = test, 
    Career = career, 
    Count = count, 
    Proportion = proportion, 
    Percent = percent
  ) %>%
  kbl(caption = "Most Common Career Chosen in Pre Test") %>%
  kable_styling()
pre_most_career
```

::: 

::: {.callout-tip icon=false collapse="true"}
## Least Common Careers in the Pre-Test
```{r}
#| label: pre-least-career-cat

# Least
pre_only_one <- pre_career %>%
  filter(count == "1") %>%
  pull(career)
cat("Careers that were chosen only once: ")
pre_only_one

pre_two_and_three <- pre_career %>%
  filter(count == 2 | count == 3) %>%
  arrange(count) %>%
  pull(career)
cat("Careers that were chosen only two or three times: ")
pre_two_and_three
```

:::

:::

::: {.callout-note icon=false collapse="true"}
## Post Test: Most and Least Common
::: {.callout-tip icon=false collapse="true"}
## Most Common Careers in the Post-Test
```{r}
#| label: post-most-career-cat

# Background
post_data_longer <- post_data %>%
  select(survey_type, career_one, career_two, career_three, career_four, career_five, career_six, career_seven, career_eight, career_nine, career_ten) %>% 
  pivot_longer(!survey_type, names_to = "career_num", values_to = "career") %>%
  select(-career_num) %>%
  filter(!is.na(career))

post_career <- post_data_longer %>%
  count(career) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "post"
  ) %>%
  select(test, career, n, proportion, percent) %>%
  rename(count = n)

# Most
post_most_career <- post_career %>%
  arrange(desc(count)) %>%
  head(10) %>%
  rename(
    Type = test, 
    Career = career, 
    Count = count, 
    Proportion = proportion, 
    Percent = percent
  ) %>%
  kbl(caption = "Most Common Career Chosen in Post Test") %>%
  kable_styling()
post_most_career
```

::: 

::: {.callout-tip icon=false collapse="true"}
## Least Common Careers in the Post-Test
```{r}
#| label: post-least-career-cat

# Least
post_only_one <- post_career %>%
  filter(count == "1") %>%
  pull(career)
cat("Careers that were chosen only once: ")
post_only_one

post_two_and_three <- post_career %>%
  filter(count == 2 | count == 3) %>%
  arrange(count) %>%
  pull(career)
cat("Careers that were chosen only two or three times: ")
post_two_and_three
```

:::

::: 

::: {.callout-note icon=false collapse="true"}
## Difference Between Pre and Post-Test Career Choices
::: {.callout-tip icon=false collapse="true"}
## Change in Careers
```{r}
#| label: diff-careers

pre_diff_careers <- pre_career[!pre_career$career %in% post_career$career,] %>%
  pull(career)
cat("New careers present in the pre-test, but not post-test: ")
pre_diff_careers

post_new_careers <- post_career[!post_career$career %in% pre_career$career,] %>%
  pull(career)
cat("Careers present in the post-test, but not pre-test: ")
post_new_careers
```

:::

::: {.callout-tip icon=false collapse="true"}
## Constant Careers
```{r}
#| label: same-careers

same_careers <- post_career[post_career$career %in% pre_career$career,] %>%
  pull(career)
cat("The same careers present in both the pre and post-tests: ")
same_careers
```

:::

::: 

## Career Major Categories

**Distribution:** It is known that students collaborated on this question at times so that's why some careers appear more often than others. However, in this section, every career option was considered only once, no matter how many students. actually wrote it down either on the Pre-Test or Post-Test. The table below groups all of the careers students wrote down by what major they would relate to in college. Out of all the 77 unique careers seen in the Pre-Test and the 56 found in the Post-Test, most careers would fall under "Medical" or "Pre-Med" majors in undergrad. 
```{r}
#| label: major-cat-compare

# Pre
pre_major_career <- data.frame (
  category  = c("Business", "Engineer", "Medical", "History", "Math", "Arts", "Physics", "Computer Science", "Legal Studies", "Earth Studies", "Design Studies", "Education", "Psych/Neuro", "Miscellaneous"),
  pre_count = c(4, 13, 26, 2, 2, 1, 6, 6, 1, 2, 2, 2, 5, 5)
) %>%
  arrange(desc(pre_count)) %>%
  mutate(pre_percent = (pre_count / sum(pre_count)) * 100)

# Post
post_major_career <- data.frame (
  category  = c("Business", "Engineer", "Medical", "History", "Math", "Arts", "Physics", "Computer Science", "Legal Studies", "Earth Studies", "Design Studies", "Education", "Psych/Neuro", "Miscellaneous"),
  post_count = c(1, 15, 20, 0, 3, 0, 3, 5, 0, 0, 0, 2, 3, 4)
) %>%
  arrange(desc(post_count)) %>%
  mutate(post_percent = (post_count / sum(post_count)) * 100)

# Compare
major_career <- full_join(pre_major_career, post_major_career) %>%
  select(category, pre_count, post_count, pre_percent, post_percent) %>%
  rename(
    Category = category, 
    `Count (Pre)` = pre_count, 
    `Count (Post)` = post_count, 
    `Percent (Pre)` = pre_percent, 
    `Precent (Post)` = post_percent
  ) %>%
  kbl(caption = "Careers Grouped by Major Category") %>%
  kable_styling()
major_career
```

## Career STEM Categories

**Distribution:** This time, the career options that were written down were grouped by which category under STE(A)M they would fall under. In the end, the majority of careers would be under the "Science" branch, being 61% of career options on the Pre-Test and 54% of the career options on the Post-Test. This is while on the Post-Test, there were no careers that would fall under the "Arts" branch even though 4% of the careers did in the Pre-Test. Besides that, the follow up for the next most common branch for the careers to fall under was "Engineering", being 17% of the careers for the Pre-Test and 27% for the Post-Test.
```{r}
#| label: stem-cat-compare

# Pre
pre_stem_cat_career <- data.frame (
  category  = c("Science", "Technology", "Engineering", "Arts", "Mathematics"),
  pre = c(47, 6, 13, 3, 8)
) %>%
  arrange(desc(pre))

# Post
post_stem_cat_career <- data.frame (
  category  = c("Science", "Technology", "Engineering", "Arts", "Mathematics"),
  post = c(30, 5, 15, 0, 6)
) %>%
  arrange(desc(post))

## Combine
stem_cat_career_all <- full_join(pre_stem_cat_career, post_stem_cat_career) %>% 
  pivot_longer(cols = c(pre, post), names_to = "Type", values_to = "Count") %>%
  group_by(Type) %>%
  mutate(percentage = Count / sum(Count) * 100) %>%
  head(9) %>%
  rename(
    Percentage = percentage, 
    Category = category
      )

stem_cat_career_all$Category <- factor(stem_cat_career_all$Category, levels = c("Science", "Technology", "Engineering", "Arts", "Mathematics"))
stem_cat_career_all$Type <- factor(stem_cat_career_all$Type, levels = c("pre", "post"))

# Compare
## Graph
colors <- c("steelblue1", "tomato")

stem_cat_plot <- ggplot(data = stem_cat_career_all, mapping = aes(x = Category, y = Percentage, fill = Type)) + 
  geom_col(position = 'dodge') + 
  geom_text(data = stem_cat_career_all, mapping = aes(x = Category, y = Percentage, label = sprintf("%.2f%%", Percentage)), position = position_dodge(width = 0.9), vjust = -0.5, size = 3) + 
  scale_fill_manual(labels = c("Pre", "Post"), values = colors) + 
  labs(
    x = "Career Category", 
    y = "Percentage", 
    title = "Careers Grouped by Their STEM Category", 
    subtitle = "Per Survey Type", 
    caption = "Source: WiSTEM 2022-2023"
  ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
stem_cat_plot
```

::: {.callout-note icon=false collapse="true"}
## Table for Careers Grouped by STEM Type
```{r}
## Table
stem_cat_career <- stem_cat_career_all %>%
  kbl(caption = "Careers Grouped by STEM Type Per Survey Type") %>%
  kable_styling()
stem_cat_career
```

:::

# Likert Scale Questions
## Question 1
**Question:** If I work hard to solve a problem, I'll find the answer. 

**Analysis:** While initially only 4.1% stated that they felt `Neutral` about their abilities to solve difficult problem if they work hard to do so, the confidence of students increased to be either `Agree` or `Strongly agree` only. Even though the percent of `Agree` decreased from 71.6% to 35.4% of all students, `Strongly agree` is now the majority at 64.6%. 
```{r}
#| label: work-hard

# Pre
pre_wh_order <- read_csv(file = "data/processed/wistem_pre_data.csv") %>%
  filter(!is.na(work_hard))
pre_wh_order$survey_type <- factor(pre_wh_order$survey_type, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

pre_wh_table <- pre_wh_order %>%
  count(work_hard) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "pre"
  )

# Post
post_wh_order <- read_csv(file = "data/processed/wistem_post_data.csv") %>%
  filter(!is.na(work_hard))
post_wh_order$work_hard <- factor(post_wh_order$work_hard, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

post_wh_table <- post_wh_order %>%
  count(work_hard) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "post"
  )

# Compare 
## Plot
percentage_data <- data %>%
  group_by(survey_type, work_hard) %>%
  summarise(count = n()) %>%
  group_by(survey_type) %>%
  mutate(percentage = count / sum(count) * 100)
percentage_data$survey_type <- factor(percentage_data$survey_type, levels = c("Pre", "Post"))

one_wh_plot <- ggplot(data = percentage_data, mapping = aes(x = work_hard, y = percentage, fill = survey_type)) + 
  geom_col(position = 'dodge') + 
  geom_text(mapping = aes(x = work_hard, y = percentage, label = sprintf("%.2f%%", percentage)), position = position_dodge(width = 0.9), vjust = -0.5, size = 3) + 
  scale_fill_manual(labels = c("Pre", "Post"), values = colors) + 
  labs(
    x = "Work Hard Response", 
    y = "Percentage", 
    title = "I'll Find the Answer if I Work Hard", 
    subtitle = "Per Survey Type", 
    caption = "Source: WiSTEM 2022-2023"
  ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

one_wh_plot
```

::: {.callout-note icon=false collapse="true"}
## Table for Work Hard
```{r}
## Table
wh_table_percent <- pre_wh_table %>%
  bind_rows(post_wh_table) %>%
  select(-n, -proportion) %>%
  pivot_wider(names_from = work_hard, values_from = percent) %>%
  kbl(caption = "Pre and Post Percent Distribution for work_hard") %>%
  kable_styling()
wh_table_percent
```

:::

## Question 2
**Question:** I am confident in my ability to apply prior knowledge to succeed in new projects. 

**Analysis:** There are very similar distributions in terms of the students' confidence in their personal knowledge to be able to handle and achieve their goals when working on tasks. This is because `Agree` remained the remained even though it decreased slightly from 66.2% to 62.5%. However, there were still some students who were `Neutral` about their capabilities. 
```{r}
#| label: knowledge

# Pre
pre_know_order <- read_csv(file = "data/processed/wistem_pre_data.csv") %>%
  filter(!is.na(knowledge))
pre_know_order$knowledge <- factor(pre_know_order$knowledge, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

pre_know_table <- pre_know_order %>%
  count(knowledge) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "pre"
  )

# Post
post_know_order <- read_csv(file = "data/processed/wistem_post_data.csv") %>%
  filter(!is.na(knowledge))
post_know_order$knowledge <- factor(post_know_order$knowledge, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

post_know_table <- post_know_order %>%
  count(knowledge) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "post"
  )

# Compare 
## Plot
percentage_data <- data %>%
  group_by(survey_type, knowledge) %>%
  summarise(count = n()) %>%
  group_by(survey_type) %>%
  mutate(percentage = count / sum(count) * 100)

one_know_plot <- ggplot(data = percentage_data, mapping = aes(x = knowledge, y = percentage, fill = survey_type)) + 
  geom_col(position = 'dodge') + 
  geom_text(mapping = aes(x = knowledge, y = percentage, label = sprintf("%.2f%%", percentage)), position = position_dodge(width = 0.9), vjust = -0.5, size = 3) + 
  scale_fill_manual(labels = c("Pre", "Post"), values = colors) + 
  labs(
    x = "Knowledge Response", 
    y = "Percentage", 
    title = "Confident in Applying Prior Knowledge to New Projects", 
    subtitle = "Per Survey Type", 
    caption = "Source: WiSTEM 2022-2023"
  ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
one_know_plot
```

::: {.callout-note icon=false collapse="true"}
## Table for Knowledge
```{r}
## Table
know_table_percent <- pre_know_table %>%
  bind_rows(post_know_table) %>%
  select(-n, -proportion) %>%
  pivot_wider(names_from = knowledge, values_from = percent) %>%
  kbl(caption = "Pre and Post Percent Distribution for knowledge") %>%
  kable_styling()
know_table_percent
```

:::

## Question 3
**Question:** I feel confident in my ability to do well in STE(A)M courses. 

**Analysis:** The distribution is generally similar with a right skew being seen and `Agree` being the majority in both the Pre and Post Test, having 54.1% and 62.5% students choosing this option, respectively. There is an improvement in as no students chose `Disagree` in terms of being confident that they can handle STE(A)M courses in general. 
```{r}
#| label: courses

# Pre
pre_course_order <- read_csv(file = "data/processed/wistem_pre_data.csv") %>%
  filter(!is.na(courses))
pre_course_order$courses <- factor(pre_course_order$courses, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

pre_course_table <- pre_course_order %>%
  count(courses) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "pre"
  )

# Post
post_course_order <- read_csv(file = "data/processed/wistem_post_data.csv") %>%
  filter(!is.na(courses))
post_course_order$courses <- factor(post_course_order$courses, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

post_course_table <- post_course_order %>%
  count(courses) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "post"
  )

# Compare
## Plot
percentage_data <- data %>%
  group_by(survey_type, courses) %>%
  summarise(count = n()) %>%
  group_by(survey_type) %>%
  mutate(percentage = count / sum(count) * 100)

one_course_plot <- ggplot(data = percentage_data, mapping = aes(x = courses, y = percentage, fill = survey_type)) + 
  geom_col(position = 'dodge') + 
  geom_text(mapping = aes(x = courses, y = percentage, label = sprintf("%.2f%%", percentage)), position = position_dodge(width = 0.9), vjust = -0.5, size = 3) + 
  scale_fill_manual(labels = c("Pre", "Post"), values = colors) + 
  labs(
    x = "Courses Response", 
    y = "Percentage", 
    title = "Confident in My Ability to do Well in STE(A)M Courses", 
    subtitle = "Per Survey Type", 
    caption = "Source: WiSTEM 2022-2023"
  ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
one_course_plot
```

::: {.callout-note icon=false collapse="true"}
## Table for Courses
```{r}
## Table
course_table_percent <- pre_course_table %>%
  bind_rows(post_course_table) %>%
  select(-n, -proportion) %>%
  pivot_wider(names_from = courses, values_from = percent) %>%
  kbl(caption = "Pre and Post Percent Distribution for courses") %>%
  kable_styling()
course_table_percent
```

:::

## Question 4
**Question:** I think I could be successful in a STE(A)M career. 

**Analysis:** The Pre Test here has a strong right skew as 1.4% of students chose `Disagree` and 1.4% chose `Strongly disagree` when considering in they could prosper in a STE(A)M field after finishing their education. There is improvement as no one chose these options in the Post Test with `Agree` remaining the majority. 
```{r}
#| label: succesful

# Pre
pre_succ_order <- read_csv(file = "data/processed/wistem_pre_data.csv") %>%
  filter(!is.na(succesful))
pre_succ_order$succesful <- factor(pre_succ_order$succesful, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

pre_succ_table <- pre_succ_order %>%
  count(succesful) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "pre"
  )

# Post
post_succ_order <- read_csv(file = "data/processed/wistem_post_data.csv") %>%
  filter(!is.na(succesful))
post_succ_order$succesful <- factor(post_succ_order$succesful, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

post_succ_table <- post_succ_order %>%
  count(succesful) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "post"
  )

# Compare
## Plot
percentage_data <- data %>%
  filter(!is.na(succesful)) %>%
  group_by(survey_type, succesful) %>%
  summarise(count = n()) %>%
  group_by(survey_type) %>%
  mutate(percentage = count / sum(count) * 100)

one_succ_plot <- ggplot(data = percentage_data, mapping = aes(x = succesful, y = percentage, fill = survey_type)) + 
  geom_col(position = 'dodge') + 
  geom_text(mapping = aes(x = succesful, y = percentage, label = sprintf("%.2f%%", percentage)), position = position_dodge(width = 0.9), vjust = -0.5, size = 3) + 
  scale_fill_manual(labels = c("Pre", "Post"), values = colors) + 
  labs(
    x = "Successful Response", 
    y = "Percentage", 
    title = "I Could be Successful in a STE(A)M Career", 
    subtitle = "Per Survey Type", 
    caption = "Source: WiSTEM 2022-2023"
  ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

one_succ_plot
```

::: {.callout-note icon=false collapse="true"}
## Table for Successful
```{r}
## Table
succ_table_percent <- pre_succ_table %>%
  bind_rows(post_succ_table) %>%
  select(-n, -proportion) %>%
  pivot_wider(names_from = succesful, values_from = percent) %>%
  kbl(caption = "Pre and Post Percent Distribution for succesful") %>%
  kable_styling()
succ_table_percent
```

:::

## Question 5
**Question:** People who work in a STE(A)M field can solve important problems. 

**Analysis:** The majority of students (59.7%) in the Pre Test are seen to `Strongly agree` with the idea that other individuals who already work in a STE(A)M field are able to solve difficult and important problems. However, 1.4% said they were `Neutral` about this. In the Post Test, there is almost an even split between students choosing either `Strongly agree` or `Agree` showing they are able to value the knowledge that people in STE(A)M have. 
```{r}
#| label: important

# Pre
pre_imp_order <- read_csv(file = "data/processed/wistem_pre_data.csv") %>%
  filter(!is.na(important))
pre_imp_order$important <- factor(pre_imp_order$important, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

pre_imp_table <- pre_imp_order %>%
  count(important) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "pre"
  )

# Post
post_imp_order <- read_csv(file = "data/processed/wistem_post_data.csv") %>%
  filter(!is.na(important))
post_imp_order$important <- factor(post_imp_order$important, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

post_imp_table <- post_imp_order %>%
  count(important) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "post"
  )

# Compare
## Plot
percentage_data <- data %>%
  filter(!is.na(important)) %>%
  group_by(survey_type, important) %>%
  summarise(count = n()) %>%
  group_by(survey_type) %>%
  mutate(percentage = count / sum(count) * 100)

one_imp_plot <- ggplot(data = percentage_data, mapping = aes(x = important, y = percentage, fill = survey_type)) + 
  geom_col(position = 'dodge') + 
  geom_text(mapping = aes(x = important, y = percentage, label = sprintf("%.2f%%", percentage)), position = position_dodge(width = 0.9), vjust = -0.5, size = 3) + 
  scale_fill_manual(labels = c("Pre", "Post"), values = colors) + 
  labs(
    x = "Important Response", 
    y = "Percentage", 
    title = "STE(A)M People Solve Important Problems", 
    subtitle = "Per Survey Type", 
    caption = "Source: WiSTEM 2022-2023"
  ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
one_imp_plot
```

::: {.callout-note icon=false collapse="true"}
## Table for Important
```{r}
## Table
imp_table_percent <- pre_imp_table %>%
  bind_rows(post_imp_table) %>%
  select(-n, -proportion) %>%
  pivot_wider(names_from = important, values_from = percent) %>%
  kbl(caption = "Pre and Post Percent Distribution for important") %>%
  kable_styling()
imp_table_percent
```

:::

## Question 6
**Question:** I am a member of a STE(A)M community. 

**Analysis:** There was an overall improvement in responses because in the pre-test, 14.3% of students chose `Disagree` while there were no students choosing this option in the post-test. This is even though there was one student who chose `Strongly disagree` in the post-test. There was also an improvement as even though `Agree` remained the majority, it increased from 38.6% to 63%. The overall goal of trying to make students feel more included in STE(A)M was achieved, but it can still be improved upon further. 
```{r}
#| label: member

# Pre
pre_mem_order <- read_csv(file = "data/processed/wistem_pre_data.csv") %>%
  filter(!is.na(member))
pre_mem_order$member <- factor(pre_mem_order$member, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

pre_mem_table <- pre_mem_order %>%
  count(member) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "pre"
  )

# Post
post_mem_order <- read_csv(file = "data/processed/wistem_post_data.csv") %>%
  filter(!is.na(member))
post_mem_order$member <- factor(post_mem_order$member, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

post_mem_table <- post_mem_order %>%
  count(member) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "post"
  )

# Compare 
## Plot
percentage_data <- data %>%
  filter(!is.na(member)) %>%
  group_by(survey_type, member) %>%
  summarise(count = n()) %>%
  group_by(survey_type) %>%
  mutate(percentage = count / sum(count) * 100)

one_mem_plot <- ggplot(data = percentage_data, mapping = aes(x = member, y = percentage, fill = survey_type)) + 
  geom_col(position = 'dodge') + 
  geom_text(mapping = aes(x = member, y = percentage, label = sprintf("%.2f%%", percentage)), position = position_dodge(width = 0.9), vjust = -0.5, size = 3) + 
  scale_fill_manual(labels = c("Pre", "Post"), values = colors) + 
  labs(
    x = "Member Response", 
    y = "Percentage", 
    title = "I'm a Member of a STE(A)M Community", 
    subtitle = "Per Survey Type", 
    caption = "Source: WiSTEM 2022-2023"
  ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
one_mem_plot
```

::: {.callout-note icon=false collapse="true"}
## Table for Member
```{r}
## Table
mem_table_percent <- pre_mem_table %>%
  bind_rows(post_mem_table) %>%
  select(-n, -proportion) %>%
  pivot_wider(names_from = member, values_from = percent) %>%
  kbl(caption = "Pre and Post Percent Distribution for member") %>%
  kable_styling()
mem_table_percent
```

:::

## Question 7
**Question:** I think a career in STE(A)M may be a good fit for me. 

**Analysis:** There was an overall improvement in responses because in the pre-test, 4.2% of students chose `Disagree` and 1.4% chose `Strongly disagree` while only 4.2% of students chose `Disagree` in the post-test. There was also an improvement as even though `Agree` remained the majority, it increased from 45.1% to 58.7% while the number of students choosing `Neutral` decreased from 19.7% to 10.9% seeing that more students are sure about a STE(A)M career fitting them in the future. 
```{r}
#| label: career

# Pre 
pre_car_order <- read_csv(file = "data/processed/wistem_pre_data.csv") %>%
  filter(!is.na(career))
pre_car_order$career <- factor(pre_car_order$career, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

pre_car_table <- pre_car_order %>%
  count(career) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "pre"
  )

# Post
post_car_order <- read_csv(file = "data/processed/wistem_post_data.csv") %>%
  filter(!is.na(career))
post_car_order$career <- factor(post_car_order$career, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

post_car_table <- post_car_order %>%
  count(career) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "post"
  )

# Compare 
## Plot
percentage_data <- data %>%
  filter(!is.na(career)) %>%
  group_by(survey_type, career) %>%
  summarise(count = n()) %>%
  group_by(survey_type) %>%
  mutate(percentage = count / sum(count) * 100)

one_car_plot <- ggplot(data = percentage_data, mapping = aes(x = career, y = percentage, fill = survey_type)) + 
  geom_col(position = 'dodge') + 
  geom_text(mapping = aes(x = career, y = percentage, label = sprintf("%.2f%%", percentage)), position = position_dodge(width = 0.9), vjust = -0.5, size = 3) + 
  scale_fill_manual(labels = c("Pre", "Post"), values = colors) + 
  labs(
    x = "Career Response", 
    y = "Percentage", 
    title = "A STE(A)M Career is a Good Fit for Me", 
    subtitle = "Per Survey Type", 
    caption = "Source: WiSTEM 2022-2023"
  ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
one_car_plot
```

::: {.callout-note icon=false collapse="true"}
## Table for Career
```{r}
## Table
car_table_percent <- pre_car_table %>%
  bind_rows(post_car_table) %>%
  select(-n, -proportion) %>%
  pivot_wider(names_from = career, values_from = percent) %>%
  kbl(caption = "Pre and Post Percent Distribution for career") %>%
  kable_styling()
car_table_percent
```

:::

## Question 8
**Question:** “Future STE(A)M professional” as part of my identity is important to me. 

**Analysis:** There was an overall improvement in responses because in the pre-test, 11.3% of students chose `Disagree` and 1.4% chose `Strongly disagree` while only 4.3% of students chose `Disagree` in the post-test. There was also an improvement as `Neutral` was the majority in the pre-test at 36.6% with `Agree` following right behind in at 35.2% meanwhile `Agree` became the majority in the post-test, having 43.5% of students choosing it. 
```{r}
#| label: identity

# Pre
pre_iden_order <- read_csv(file = "data/processed/wistem_pre_data.csv") %>%
  filter(!is.na(identity))
pre_iden_order$identity <- factor(pre_iden_order$identity, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

pre_iden_table <- pre_iden_order %>%
  count(identity) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "pre"
  )

# Post
post_iden_order <- read_csv(file = "data/processed/wistem_post_data.csv") %>%
  filter(!is.na(identity))
post_iden_order$identity <- factor(post_iden_order$identity, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

post_iden_table <- post_iden_order %>%
  count(identity) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "post"
  )

# Compare 
## Plot
percentage_data <- data %>%
  filter(!is.na(identity)) %>%
  group_by(survey_type, identity) %>%
  summarise(count = n()) %>%
  group_by(survey_type) %>%
  mutate(percentage = count / sum(count) * 100)

one_iden_plot <- ggplot(data = percentage_data, mapping = aes(x = identity, y = percentage, fill = survey_type)) + 
  geom_col(position = 'dodge') + 
  geom_text(mapping = aes(x = identity, y = percentage, label = sprintf("%.2f%%", percentage)), position = position_dodge(width = 0.9), vjust = -0.5, size = 3) + 
  scale_fill_manual(labels = c("Pre", "Post"), values = colors) + 
  labs(
    x = "Identity Response", 
    y = "Percentage", 
    title = "`Future STE(A)M Professional` is a Part of My Identity", 
    subtitle = "Per Survey Type", 
    caption = "Source: WiSTEM 2022-2023"
  ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
one_iden_plot
```

::: {.callout-note icon=false collapse="true"}
## Table for Identity
```{r}
## Table
iden_table_percent <- pre_iden_table %>%
  bind_rows(post_iden_table) %>%
  select(-n, -proportion) %>%
  pivot_wider(names_from = identity, values_from = percent) %>%
  kbl(caption = "Pre and Post Percent Distribution for identity") %>%
  kable_styling()
iden_table_percent
```

:::

## Question 9
**Question:** If I entered a STE(A)M career, I could make a meaningful contribution to society. 

**Analysis:** There was an overall improvement in responses because in the pre-test, 2.8% of students chose `Disagree` while none chose this option in the post-test. There was also an improvement as even though `Agree` remained the majority, `Strongly agree` was chosen 32.6% of the time in the post-test compared to that of the pre-test. Therefore, there is an increase in students feeling that they themselves can make good change if they were to have a STE(A)M career.  
```{r}
#| label: meaningful

# Pre 
pre_mean_order <- read_csv(file = "data/processed/wistem_pre_data.csv") %>%
  filter(!is.na(meaningful))
pre_mean_order$meaningful <- factor(pre_mean_order$meaningful, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

pre_mean_table <- pre_mean_order %>%
  count(meaningful) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "pre"
  )

# Post 
post_mean_order <- read_csv(file = "data/processed/wistem_post_data.csv") %>%
  filter(!is.na(meaningful))
post_mean_order$meaningful <- factor(post_mean_order$meaningful, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

post_mean_table <- post_mean_order %>%
  count(meaningful) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "post"
  )

# Compare 
## Plot
percentage_data <- data %>%
  filter(!is.na(meaningful)) %>%
  group_by(survey_type, meaningful) %>%
  summarise(count = n()) %>%
  group_by(survey_type) %>%
  mutate(percentage = count / sum(count) * 100)

one_mean_plot <- ggplot(data = percentage_data, mapping = aes(x = meaningful, y = percentage, fill = survey_type)) + 
  geom_col(position = 'dodge') + 
  geom_text(mapping = aes(x = meaningful, y = percentage, label = sprintf("%.2f%%", percentage)), position = position_dodge(width = 0.9), vjust = -0.5, size = 3) + 
  scale_fill_manual(labels = c("Pre", "Post"), values = colors) + 
  labs(
    x = "Meaningful Response", 
    y = "Percentage", 
    title = "I Would Make Meaningful Contribution in STE(A)M",
    subtitle = "Per Survey Type", 
    caption = "Source: WiSTEM 2022-2023"
  ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
one_mean_plot
```

::: {.callout-note icon=false collapse="true"}
## Table for Meaningful
```{r}
## Table
mean_table_percent <- pre_mean_table %>%
  bind_rows(post_mean_table) %>%
  select(-n, -proportion) %>%
  pivot_wider(names_from = meaningful, values_from = percent) %>%
  kbl(caption = "Pre and Post Percent Distribution for meaningful") %>%
  kable_styling()
mean_table_percent
```

:::
## Question 10
**Question:** I understand what everyday STE(A)M work is like. 

**Analysis:** The distribution for the choice options are pretty similar between the pre-test and post-test, but the improvements are that `Disagree` decreased from 11.4% to 4.3% while `Strongly agree` increased from 4.3% to 17.4%. This is while `Agree` remained the majority in both scenarios. Therefore, students already had a good understanding on what STE(A)M work looked like, but they were still able to learn new information about it. 
```{r}
#| label: everyday

# Pre 
pre_every_order <- read_csv(file = "data/processed/wistem_pre_data.csv") %>%
  filter(!is.na(everyday))
pre_every_order$everyday <- factor(pre_every_order$everyday, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

pre_every_table <- pre_every_order %>%
  count(everyday) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "pre"
  )

# Post
post_every_order <- read_csv(file = "data/processed/wistem_post_data.csv") %>%
  filter(!is.na(everyday))
post_every_order$everyday <- factor(post_every_order$everyday, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

post_every_table <- post_every_order %>%
  count(everyday) %>%
  mutate(
    proportion = n / sum(n), 
    percent = proportion * 100, 
    test = "post"
  )

# Compare 
## Plot
percentage_data <- data %>%
  filter(!is.na(everyday)) %>%
  group_by(survey_type, everyday) %>%
  summarise(count = n()) %>%
  group_by(survey_type) %>%
  mutate(percentage = count / sum(count) * 100)

one_every_plot <- ggplot(data = percentage_data, mapping = aes(x = everyday, y = percentage, fill = survey_type)) + 
  geom_col(position = 'dodge') + 
  geom_text(mapping = aes(x = everyday, y = percentage, label = sprintf("%.2f%%", percentage)), position = position_dodge(width = 0.9), vjust = -0.5, size = 3) + 
  scale_fill_manual(labels = c("Pre", "Post"), values = colors) + 
  labs(
    x = "Everday Response", 
    y = "Percentage", 
    title = "I Understand Everyday STE(A)M Work", 
    subtitle = "Per Survey Type", 
    caption = "Source: WiSTEM 2022-2023"
  ) + 
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
one_every_plot
```

::: {.callout-note icon=false collapse="true"}
## Table for Everyday
```{r}
## Table
every_table_percent <- pre_every_table %>%
  bind_rows(post_every_table) %>%
  select(-n, -proportion) %>%
  pivot_wider(names_from = everyday, values_from = percent) %>%
  kbl(caption = "Pre and Post Percent Distribution for everyday") %>%
  kable_styling()
every_table_percent
```

:::
