==> 02_wistem_exit_slips.qmd <==
---
title: "WiSTEM 2022 - 2023"
subtitle: "Exit Slip Analyses"

format:
  html:
    toc: true
    embed-resources: true
    echo: false
    link-external-newwindow: true
execute:
  warning: false
  message: false

from: markdown+emoji  
editor: 
  markdown: 
    wrap: 72
---

```{r}
#| label: load-packages

library(tidyverse)
library(patchwork)
library(readr)
library(kableExtra)

set.seed(1968)
```

```{r}
#| label: load-data-set

data <- read_csv(file = "data/processed/wistem_exit_slips_clean.csv") %>%
  filter(!is.na(interesting), !is.na(included), !is.na(belong))

data$interesting <- factor(data$interesting, levels = c("Somewhat", "Fairly", "Very"))
data$included <- factor(data$included, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))
data$belong <- factor(data$belong, levels = c("Strongly disagree", "Neutral", "Agree", "Strongly agree"))
```

# Background

**WiSTEM:** In total, there were 17 overall WiSTEM meetings held throughout the 2022-2023 academic school year. However, there were 13 meetings with attendance data and 14 meetings with exit slip data. Some of these overlap in missing both attendance and exit slip data.

**Exit Slips Distribution:** Overall, there were 483 total exit slips collected over the course of the 2022-2023 academic school year. The graph and table below show the number of exit slips obtained for each of the 14 meetings where exit slips were provided: 
```{r}
#| label: exit-slip-per-meeting

# Number per meeting
date_order <- read_csv(file = "data/processed/wistem_exit_slips_clean.csv") %>%
  filter(!is.na(date))
date_order$date <- factor(date_order$date, levels = c("May 31st", "May 12th", "Apr 14th", "Mar 31st", "Mar 3rd", "Feb 24th", "Feb 10th", "Jan 27th", "Dec 16th", "Dec 2nd", "Nov 18th", "Oct 28th", "Oct 14th", "Sep 30th"))

# Graph
ggplot(data = date_order, mapping = aes(y = date)) + 
  geom_bar(color = "azure4", fill = "gainsboro") + 
  geom_text(stat='count', aes(label=..count..), position = position_stack(vjust = 0.9)) +
  labs(
    title = "Number of Exit Slips Per Meeting", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Number of Students", 
    y = "Meeting Date"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
```

::: {.callout-note icon=false collapse="true"}
## Table for Number of Exit Slips Per Meeting
```{r}
# Table
date_order %>%
  filter(!is.na(date)) %>%
  count(date) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  rename(count = n) %>%
  kbl(caption = "Number of Exit Slips Per Meeting") %>%
  kable_styling()
```
:::

# Completion Rate

**Distribution:** Even though there were 13 meetings with attendance data, only 12 meetings overlap with the meetings that also have exit slip data collected throughout the 2022-2023 school year. Overall, the majority of meetings have an over 70% exit slip completion rate based on the number of students that attended the meeting. However, 3 meetings have completion rates of less than 50%. Besides that, the `Sep 30th` meeting seems to have an issue as there were more exit slips collected than the number of students that signed their name on the attendance sheet. 
```{r}
#| label: completion

attendance_data <- read_csv(file = "data/processed/wistem_attendance_clean.csv") %>%
  filter(!is.na(date))
attendance_data$date <- factor(attendance_data$date, levels = c("May 31st", "May 12th", "Apr 14th", "Mar 31st", "Mar 3rd", "Feb 24th", "Feb 10th", "Jan 27th", "Dec 16th", "Dec 2nd", "Nov 18th", "Oct 28th", "Oct 14th", "Sep 30th"))

attendance_data <- attendance_data %>%
  filter(!is.na(date))

# Table
completion_rate <- data.frame (
  date  = c("May 31st", "May 12th", "Apr 14th", "Mar 31st", "Mar 3rd", "Feb 24th", "Feb 10th", "Jan 27th", "Dec 16th", "Dec 2nd", "Nov 18th", "Oct 28th", "Oct 14th", "Sep 30th"),
  exit_slips = c(7, 43, 35, 40, 41, 39, 24, 39, 41, 43, 21, 18, 29, 64), 
  attendance = c(NA, 46, 40, 57, 51, 48, 51, 44, 54, 48, 51, 48, 35, 48)
) %>%
  mutate(
    completion_prop = exit_slips / attendance, 
    completion_percent = completion_prop * 100
    ) %>%
  rename(
    Date = date, 
    `Exit Slip Count` = exit_slips, 
    `Attendance Count` = attendance, 
    Proportion = completion_prop, 
    Percent = completion_percent
  )

completion_rate %>%
  kbl(caption = "Exit Slip Completion Rate Per Meeting") %>%
  kable_styling()
```

# Overall Distribution Per Category
## Question 1
**Question:** How interesting was today's meeting?

**Distribution:** Overall, throughout the 2022-2023 academic school year, a majority (64.8%) of students found the meeting topics to be `Very` interesting, considering all possible meeting topics of Activity, Discussion, Guest Speaker, Out of School Event, and Guest Speaker / Discussion. 
```{r}
#| label: avg-interesting

# Graph
interesting_order <- read_csv(file = "data/processed/wistem_exit_slips_clean.csv") %>%
  filter(!is.na(interesting))
interesting_order$interesting <- factor(interesting_order$interesting, levels = c("Somewhat", "Fairly", "Very"))

percentage_data <- interesting_order %>%
  group_by(interesting) %>%
  summarize(percentage = n() / nrow(interesting_order) * 100)

ggplot(data = interesting_order, mapping = aes(x = interesting)) + 
  geom_bar(color = "azure4", fill = "gainsboro") + 
  geom_text(data = percentage_data, aes(x = interesting, y = percentage, label = paste0(round(percentage, 1), "%")),
            position = position_dodge(width = 0.9), vjust = -0.5) +
  labs(
    title = "How Interesting Was Today’s Meeting?", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Number of Students", 
    y = "Meeting Date"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
```

::: {.callout-note icon=false collapse="true"}
## Table for Overall Interesting
```{r}
# Table
data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Distribution for Interesting") %>%
  kable_styling()
```
:::

## Question 2
**Question:** I feel included in WiSTEM. 

**Distribution:** Overall, throughout the 2022-2023 academic school year, a majority (63.8%) of students said they `Strongly agree` with the statement that they feel included in WiSTEM, not considering the general trend from the start of the school year to the end. Granted, 0.8% and 0.2% of students throughout the school year (with repeats) chose `Strongly disagree` and `Disagree`, respectively, but this was only a total of 5 student choices.
```{r}
#| label: avg-included

# Graph
included_order <- read_csv(file = "data/processed/wistem_exit_slips_clean.csv") %>%
  filter(!is.na(included))
included_order$included <- factor(included_order$included, levels = c("Strongly disagree", "Disagree", "Neutral", "Agree", "Strongly agree"))

percentage_data <- included_order %>%
  group_by(included) %>%
  summarize(percentage = n() / nrow(included_order) * 100)

ggplot(data = included_order, mapping = aes(x = included)) + 
  geom_bar(color = "azure4", fill = "gainsboro") + 
  geom_text(data = percentage_data, aes(x = included, y = percentage, label = paste0(round(percentage, 1), "%")),
            position = position_dodge(width = 0.9), vjust = -0.5) + 
  labs(
    title = "I Feel Included in WiSTEM", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Number of Students", 
    y = "Meeting Date"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
```

::: {.callout-note icon=false collapse="true"}
## Table for Overall Included
```{r}
# Table
data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Distribution for Included") %>%
  kable_styling()
```
:::

## Question 3
**Question:** I belong in a STE(A)M community. 

**Distribution:** Overall, throughout the 2022-2023 academic school year, a majority (59.9%) of students said they `Strongly agree` with the statement that they belong in the STE(A)M world, not considering the general trend from the start of the school year to the end. Granted, 0.8% of students throughout the school year (with repeats) chose `Strongly disagree`, but this was only a total of 4 student choices.  
```{r}
#| label: avg-belong

# Graph
belong_order <- read_csv(file = "data/processed/wistem_exit_slips_clean.csv") %>%
  filter(!is.na(belong))
belong_order$belong <- factor(belong_order$belong, levels = c("Strongly disagree", "Neutral", "Agree", "Strongly agree"))

percentage_data <- belong_order %>%
  group_by(belong) %>%
  summarize(percentage = n() / nrow(belong_order) * 100)

ggplot(data = belong_order, mapping = aes(x = belong)) + 
  geom_bar(color = "azure4", fill = "gainsboro") + 
  geom_text(data = percentage_data, aes(x = belong, y = percentage, label = paste0(round(percentage, 1), "%")),
            position = position_dodge(width = 0.9), vjust = -0.5) +  
  labs(
    title = "I Belong in a STE(A)M Community", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Number of Students", 
    y = "Meeting Date"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))
```

::: {.callout-note icon=false collapse="true"}
## Table for Overall Belong
```{r}
# Table
data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Distribution for belong") %>%
  kable_styling()
```
:::

# Distribution Per Meeting

## September 30, 2022
**Topic:** Discussion

**Distribution:** Being the first WiSTEM meeting of the year, a majority (70.3%) of the students found the meeting to be `Fairly` interesting. This trend is also seen with a majority (50%) saying they `Agree` with the statement that they feel included in WiSTEM and a majority (50%) saying they `Agree` with the statement that they feel that they belong in the STE(A)M community. Throughout all three questions, there is a similar distribution. 
```{r}
#| label: sep-30

sep_30_data <- data %>%
  filter(date == "Sep 30th")

# interesting
sep_30_interesting <- sep_30_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = sep_30_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "September 30th Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))


sep_30_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Sep 30th Distribution for interesting") %>%
  kable_styling()

# included
sep_30_included <- sep_30_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = sep_30_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "September 30th Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

sep_30_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Sep 30th Distribution for included") %>%
  kable_styling()

# belong
sep_30_belong <- sep_30_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = sep_30_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "September 30th Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

sep_30_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Sep 30th Distribution for belong") %>%
  kable_styling()
```

## October 14, 2022
**Topic:** Activity

**Distribution:** A majority (58.6%) of the students found the meeting to be `Very` interesting, having no students choosing the `Somewhat` option. This trend is also seen with a majority (55.2%) saying they `Strongly agree` with the statement that they feel included in WiSTEM and a majority (51.7%) saying they `Strongly agree` with the statement that they feel that they belong in the STE(A)M community. Therefore, from meeting one to meeting two, there seemed to be an improvement in terms of a sense of community being built in the WiSTEM club space.  
```{r}
#| label: oct-14

oct_14_data <- data %>%
  filter(date == "Oct 14th")

# interesting
oct_14_interesting <- oct_14_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = oct_14_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "October 14th Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

oct_14_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Oct 14th Distribution for interesting") %>%
  kable_styling()

# included
oct_14_included <- oct_14_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = oct_14_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "October 14th Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

oct_14_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Oct 14th Distribution for included") %>%
  kable_styling()

# belong
oct_14_belong <- oct_14_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = oct_14_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "October 14th Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

oct_14_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Oct 14th Distribution for belong") %>%
  kable_styling()
```

## October 28, 2022
**Topic:** Activity

**Distribution:** A majority (50%) of the students found the meeting to be `Fairly` interesting. Even though 44.4% of students thought it was `Very` interesting, it is slightly surprising that this activity based meeting was not more engaging as the meeting before this. However, the same trend can be seen from the last meeting with a majority (55.6%) saying they `Strongly agree` with the statement that they feel included in WiSTEM and a majority (50%) saying they `Strongly agree` with the statement that they feel that they belong in the STE(A)M community. Therefore, a sense of community that is being built has started off strong and remains strong into the third meeting.
```{r}
#| label: oct-28

oct_28_data <- data %>%
  filter(date == "Oct 28th")

# interesting
oct_28_interesting <- oct_28_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = oct_28_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "October 28th Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

oct_28_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Oct 28th Distribution for interesting") %>%
  kable_styling()

# included
oct_28_included <- oct_28_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = oct_28_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "October 28th Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

oct_28_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Oct 28th Distribution for included") %>%
  kable_styling()

# belong
oct_28_belong <- oct_28_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = oct_28_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "October 28th Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

oct_28_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Oct 28th Distribution for belong") %>%
  kable_styling()
```

## November 18, 2022
**Topic:** Activity

**Distribution:** A majority (57.1%) of the students found the meeting to be `Very` interesting, having no students choosing the `Somewhat` option. However, a majority (57.1%) said they `Agree` with the statement that they feel included in WiSTEM and a majority (52.4%) said they `Agree` with the statement that they feel that they belong in the STE(A)M community. Therefore, there seemed to be a slight decrease in a sense of community, especially since 1 student said they `Strongly disagree` with both statements.   
```{r}
#| label: nov-18

nov_18_data <- data %>%
  filter(date == "Nov 18th")

# interesting
nov_18_interesting <- nov_18_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = nov_18_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "November 18th Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

nov_18_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Nov 18th Distribution for interesting") %>%
  kable_styling()

# included
nov_18_included <- nov_18_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = nov_18_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "November 18th Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

nov_18_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Nov 18th Distribution for included") %>%
  kable_styling()

# belong
nov_18_belong <- nov_18_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = nov_18_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "November 18th Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

nov_18_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Nov 18th Distribution for belong") %>%
  kable_styling()
```

## December 2, 2022
**Topic:** Discussion

**Distribution:** Even though this meeting was discussion based, a majority (62.8%) of the students found the meeting to be `Very` interesting. The same trend from the first couple of meetings can be seen with a majority (60.5%) saying they `Strongly agree` with the statement that they feel included in WiSTEM and a majority (48.8%) saying they `Strongly agree` with the statement that they feel that they belong in the STE(A)M community. However, there was one student who, again, said they `Strongly disagree` with both statements. 
```{r}
#| label: dec-2

dec_2_data <- data %>%
  filter(date == "Dec 2nd")

# interesting
dec_2_interesting <- dec_2_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = dec_2_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "December 2nd Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

dec_2_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Dec 2nd Distribution for interesting") %>%
  kable_styling()

# included
dec_2_included <- dec_2_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = dec_2_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "December 2nd Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

dec_2_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Dec 2nd Distribution for included") %>%
  kable_styling()

# belong
dec_2_belong <- dec_2_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = dec_2_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "December 2nd Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

dec_2_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Dec 2nd Distribution for belong") %>%
  kable_styling()
```

## December 16, 2022
**Topic:** Activity

**Distribution:** An overwhelming majority (92.7%) of the students found the meeting to be `Very` interesting. The same trend can be seen from the last meeting with a majority (80.5%) saying they `Strongly agree` with the statement that they feel included in WiSTEM and a majority (75.6%) saying they `Strongly agree` with the statement that they feel that they belong in the STE(A)M community. However, there was one student who, again, said they `Strongly disagree` with both statements. 
```{r}
#| label: dec-16

dec_16_data <- data %>%
  filter(date == "Dec 16th")

# interesting
dec_16_interesting <- dec_16_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = dec_16_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "December 16th Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

dec_16_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Dec 16th Distribution for interesting") %>%
  kable_styling()

# included
dec_16_included <- dec_16_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = dec_16_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "December 16th Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

dec_16_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Dec 16th Distribution for included") %>%
  kable_styling()

# belong
dec_16_belong <- dec_16_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = dec_16_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "December 16th Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

dec_16_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Dec 16th Distribution for belong") %>%
  kable_styling()
```

## January 27, 2023
**Topic:** Activity

**Distribution:** A majority (56.4%) of the students found the meeting to be `Very` interesting. The same trend can be seen from the last meeting with a majority (66.7%) saying they `Strongly agree` with the statement that they feel included in WiSTEM  and a majority (69.2%) saying they `Strongly agree` with the statement that they feel that they belong in the STE(A)M community. In both cases, no student chose an option below `Agree`. 
```{r}
#| label: jan-27

jan_27_data <- data %>%
  filter(date == "Jan 27th")

# interesting
jan_27_interesting <- jan_27_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = jan_27_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "January 27th Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

jan_27_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Jan 27th Distribution for interesting") %>%
  kable_styling()

# included
jan_27_included <- jan_27_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = jan_27_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "January 27th Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

jan_27_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Jan 27th Distribution for included") %>%
  kable_styling()

# belong
jan_27_belong <- jan_27_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = jan_27_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "January 27th Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

jan_27_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Jan 27th Distribution for belong") %>%
  kable_styling()
```

## February 10, 2023
**Topic:** Activity

**Distribution:** An overwhelming majority (91.7%) of the students found the meeting to be `Very` interesting. The same trend can be seen from the last meeting with a majority (83.3%) saying they `Strongly agree` with the statement that they feel included in WiSTEM  and a majority (83.3%) saying they `Strongly agree` with the statement that they feel that they belong in the STE(A)M community. In the second question, no student chose an option below `Agree` while only one student chose `Neutral` in the third question. Therefore, there was still a strong sense of community being developed. 
```{r}
#| label: feb-10

feb_10_data <- data %>%
  filter(date == "Feb 10th")

# interesting
feb_10_interesting <- feb_10_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = feb_10_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "February 10th Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

feb_10_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Feb 10th Distribution for interesting") %>%
  kable_styling()

# included
feb_10_included <- feb_10_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = feb_10_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "February 10th Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

feb_10_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Feb 10th Distribution for included") %>%
  kable_styling()

# belong
feb_10_belong <- feb_10_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = feb_10_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "February 10th Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

feb_10_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Feb 10th Distribution for belong") %>%
  kable_styling()
```

## February 24, 2023
**Topic:** Discussion

**Distribution:** A majority (53.8%) of the students found the meeting to be `Very` interesting. The same trend can be seen from the last meeting with a majority (64.1%) saying they `Strongly agree` with the statement that they feel included in WiSTEM  and a majority (56.4%) saying they `Strongly agree` with the statement that they feel that they belong in the STE(A)M community. In both cases, the lowest option being chosen was `Neutral`. 
```{r}
#| label: feb-24

feb_24_data <- data %>%
  filter(date == "Feb 24th")

# interesting
feb_24_interesting <- feb_24_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = feb_24_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "February 24th Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

feb_24_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Feb 24th Distribution for interesting") %>%
  kable_styling()

# included
feb_24_included <- feb_24_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = feb_24_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "February 24th Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

feb_24_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Feb 24th Distribution for included") %>%
  kable_styling()

# belong
feb_24_belong <- feb_24_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = feb_24_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "February 24th Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

feb_24_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Feb 24th Distribution for belong") %>%
  kable_styling()
```

## March 3, 2023
**Topic:** Guest Speaker

**Distribution:** A majority (78%) of the students found the meeting to be `Very` interesting, having no students choosing the `Somewhat` option. The same trend can be seen from the last meeting with a majority (80.5%) saying they `Strongly agree` with the statement that they feel included in WiSTEM  and a majority (70.7%) saying they `Strongly agree` with the statement that they feel that they belong in the STE(A)M community. In the second question, only one student chose `Neutral` in the third question while no student chose an option below `Agree` in the third question. Therefore, there was still a strong sense of community being developed. 
```{r}
#| label: mar-3

mar_3_data <- data %>%
  filter(date == "Mar 3rd")

# interesting
mar_3_interesting <- mar_3_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = mar_3_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "March 3rd Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

mar_3_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Mar 3rd Distribution for interesting") %>%
  kable_styling()

# included
mar_3_included <- mar_3_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = mar_3_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "March 3rd Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

mar_3_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Mar 3rd Distribution for included") %>%
  kable_styling()

# belong
mar_3_belong <- mar_3_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = mar_3_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "March 3rd Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

mar_3_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Mar 3rd Distribution for belong") %>%
  kable_styling()
```

## March 31, 2023
**Topic:** Activity

**Distribution:** A majority (76.9%) of the students found the meeting to be `Very` interesting, having no students choosing the `Somewhat` option. The same trend can be seen from the last meeting with a majority (79.5%) saying they `Strongly agree` with the statement that they feel included in WiSTEM  and a majority (66.7%) saying they `Strongly agree` with the statement that they feel that they belong in the STE(A)M community. In both cases, no student chose an option below `Agree`. 
```{r}
#| label: mar-31

mar_31_data <- data %>%
  filter(date == "Mar 31st")

# interesting
mar_31_interesting <- mar_31_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = mar_31_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "March 31st Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

mar_31_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Mar 31st Distribution for interesting") %>%
  kable_styling()

# included
mar_31_included <- mar_31_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = mar_31_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "March 31st Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

mar_31_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Mar 31st Distribution for included") %>%
  kable_styling()

# belong
mar_31_belong <- mar_31_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = mar_31_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "March 31st Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

mar_31_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Mar 31st Distribution for belong") %>%
  kable_styling()
```

## April 14, 2023
**Topic:** Activity

**Distribution:** An overwhelming majority (94.3%) of the students found the meeting to be `Very` interesting. The same trend can be seen from the last meeting with a majority (77.1%) saying they `Strongly agree` with the statement that they feel included in WiSTEM  and a majority (71.4%) saying they `Strongly agree` with the statement that they feel that they belong in the STE(A)M community. However, there was one student who, again after a couple of meeting, saying they `Strongly disagree` with both statements. 
```{r}
#| label: apr-14

apr_14_data <- data %>%
  filter(date == "Apr 14th")

# interesting
apr_14_interesting <- apr_14_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = apr_14_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "April 14th Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

apr_14_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Apr 14th Distribution for interesting") %>%
  kable_styling()

# included
apr_14_included <- apr_14_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = apr_14_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "April 14th Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

apr_14_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Apr 14th Distribution for included") %>%
  kable_styling()

# belong
apr_14_belong <- apr_14_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = apr_14_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "April 14th Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

apr_14_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Apr 14th Distribution for belong") %>%
  kable_styling()
```

## May 12, 2023
**Topic:** Guest Speaker / Discussion

**Distribution:** A majority (72.1%) of the students found the meeting to be `Very` interesting even though it was a different meeting topic. The same trend can be seen from the last meeting with a majority (76.7%) saying they `Strongly agree` with the statement that they feel included in WiSTEM and the same majority (76.7%) saying they `Strongly agree` with the statement that they feel that they belong in the STE(A)M community. In both cases, no student chose an option below `Agree`.
```{r}
#| label: may-12

may_12_data <- data %>%
  filter(date == "May 12th")

# interesting
may_12_interesting <- may_12_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = may_12_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "May 12th Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

may_12_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "May 12th Distribution for interesting") %>%
  kable_styling()

# included
may_12_included <- may_12_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = may_12_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "May 12th Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

may_12_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "May 12th Distribution for included") %>%
  kable_styling()

# belong
may_12_belong <- may_12_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = may_12_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "May 12th Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

may_12_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "May 12th Distribution for belong") %>%
  kable_styling()
```

## May 31, 2023
**Topic:** Out of School Event

**Distribution:** An overwhelming majority (85.7%) of the students found the meeting to be `Very` interesting even though it was a different meeting topic. The same trend can be seen from the last meeting with a majority (71.4%) saying they `Strongly agree` with the statement that they feel included in WiSTEM and the same majority (71.4%) saying they `Strongly agree` with the statement that they feel that they belong in the STE(A)M community. However, there was one student who, again after a couple of meeting, saying they `Strongly disagree` with both statements. Overall, there was a stronger sense of community that was being built through the use of WiSTEM. 
```{r}
#| label: may-31

may_31_data <- data %>%
  filter(date == "May 31st")

# interesting
may_31_interesting <- may_31_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = may_31_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "May 31st Interesting Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

may_31_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "May 31st Distribution for interesting") %>%
  kable_styling()

# included
may_31_included <- may_31_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = may_31_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "May 31st Included Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

may_31_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "May 31st Distribution for included") %>%
  kable_styling()

# belong
may_31_belong <- may_31_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = may_31_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "May 31st Belong Distribution", 
    subtitle = "Unique Students", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

may_31_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "May 31st Distribution for belong") %>%
  kable_styling()
```

# Distribution Per Topic

## Activity

**Distribution:** There are 8 meetings that had a topic of `Activity`, accounting for 8/14 or 57.14% of the exit slip data collected for the 2022-2023 school year. Overall, throughout all of the meetings, 70.3% of students (with repeats) though the Activity based meetings were `Very` interesting. Additionally, throughout all of the Activity based meetings (which were the majority of the meetings), 66.3% of students (with repeats) stated that they `Strongly agree` with the statement that they feel included in WiSTEM while 63% stated they `Strongly agree` with the statement that they belong in the STE(A)M community. 
```{r}
#| label: activity-distribution

activity_data <- data %>%
  filter(topic == "Activity")

# interesting
activity_interesting <- activity_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = activity_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Activity Interesting Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

activity_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Activity Distribution for interesting") %>%
  kable_styling()

# included
activity_included <- activity_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = activity_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Activity Included Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

activity_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Activity Distribution for included") %>%
  kable_styling()

# belong
activity_belong <- activity_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = activity_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Activity Belong Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

activity_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Activity Distribution for belong") %>%
  kable_styling()
```

## Discussion

**Distribution:** There are 2 meetings that had a topic of `Discussion`, accounting for 2/14 or 14.29% of the exit slip data collected for the 2022-2023 school year. Overall, throughout all of the meetings, 55.1% of students (with repeats) though the Discussion based meetings were `Fairly` interesting. Additionally, throughout all of the Discussion based meetings, 50.5% of students (with repeats) stated that they `Agree` with the statement that they feel included in WiSTEM while 46.7% stated they `Agree` with the statement that they belong in the STE(A)M community. 
```{r}
#| label: discussion-distribution

discussion_data <- data %>%
  filter(topic == "Discussion")

# interesting
discussion_interesting <- discussion_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = discussion_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Discussion Interesting Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

discussion_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Discussion Distribution for interesting") %>%
  kable_styling()

# included
discussion_included <- discussion_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = discussion_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Discussion Included Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

discussion_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Discussion Distribution for included") %>%
  kable_styling()

# belong
discussion_belong <- discussion_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = discussion_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Discussion Belong Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

discussion_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Discussion Distribution for belong") %>%
  kable_styling()
```

## Guest Speaker

**Distribution:** There are 2 meetings that had a topic of `Guest Speaker`, accounting for 2/14 or 14.29% of the exit slip data collected for the 2022-2023 school year. Overall, throughout all of the meetings, 77.5% of students (with repeats) though the Guest Speaker based meetings were `Very` interesting with no one choose the `Somewhat` option. Additionally, throughout all of the Guest Speaker based meetings, 80% of students (with repeats) stated that they `Strongly agree` with the statement that they feel included in WiSTEM while 68.8% stated they `Strongly agree` with the statement that they belong in the STE(A)M community. 
```{r}
#| label: guest-distribution

guest_data <- data %>%
  filter(topic == "Guest Speaker")

# interesting
guest_interesting <- guest_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = guest_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Guest Speaker Interesting Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

guest_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Guest Speaker Distribution for interesting") %>%
  kable_styling()

# included
guest_included <- guest_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = guest_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Guest Speaker Included Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

guest_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Guest Speaker Distribution for included") %>%
  kable_styling()

# belong
guest_belong <- guest_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = guest_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Guest Speaker Belong Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

guest_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Guest Speaker Distribution for belong") %>%
  kable_styling()
```

## Guest Speaker / Discussion

**Distribution:** There are 1 meeting that had a topic of `Guest / Discussion`, accounting for 1/14 or 7.14% of the exit slip data collected for the 2022-2023 school year. Overall, throughout all of the meetings, 72.1% of students (with repeats) though the Guest Speaker / Discussion based meetings were `Very` interesting with no one choose the `Somewhat` option. Additionally, throughout all of the Guest Speaker / Discussion based meetings, 76.7% of students (with repeats) stated that they `Strongly agree` with the statement that they feel included in WiSTEM while 76.7% stated they `Strongly agree` with the statement that they belong in the STE(A)M community. 
```{r}
#| label: guest-discussion-distribution

guest_disc_data <- data %>%
  filter(topic == "Guest / Discussion")

# interesting
guest_disc_interesting <- guest_disc_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = guest_disc_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Guest / Discussion Interesting Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

guest_disc_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Guest / Discussion Distribution for interesting") %>%
  kable_styling()

# included
guest_disc_included <- guest_disc_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = guest_disc_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Guest / Discussion Included Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

guest_disc_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Guest / Discussion Distribution for included") %>%
  kable_styling()

# belong
guest_disc_belong <- guest_disc_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = guest_disc_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Guest / Discussion Belong Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

guest_disc_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Guest / Discussion Distribution for belong") %>%
  kable_styling()
```

## Out of School Event

**Distribution:** There are 1 meeting that had a topic of `Out of School Event`, accounting for 1/14 or 7.14% of the exit slip data collected for the 2022-2023 school year. Overall, throughout all of the meetings, 85.7% of students (with repeats) though the Out of School Event based meetings were `Very` interesting with no one choose the `Somewhat` option. Additionally, throughout all of the Out of School Event based meetings, 71.4% of students (with repeats) stated that they `Strongly agree` with the statement that they feel included in WiSTEM while 71.4% stated they `Strongly agree` with the statement that they belong in the STE(A)M community. 
```{r}
#| label: out-of-school-distribution

oose_data <- data %>%
  filter(topic == "Out of School Activity")

# interesting
oose_interesting <- oose_data %>%
  group_by(interesting) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = oose_interesting, mapping = aes(x = interesting, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Out of School Event Interesting Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Interesting Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

oose_data %>%
  filter(!is.na(interesting)) %>%
  count(interesting) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Out of School Activity Distribution for interesting") %>%
  kable_styling()

# included
oose_included <- oose_data %>%
  group_by(included) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = oose_included, mapping = aes(x = included, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Out of School Event Included Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Included Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

oose_data %>%
  filter(!is.na(included)) %>%
  count(included) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Out of School Activity Distribution for included") %>%
  kable_styling()

# belong
oose_belong <- oose_data %>%
  group_by(belong) %>%
  summarise(count = n()) %>%
  mutate(percentage = count / sum(count) * 100)

ggplot(data = oose_belong, mapping = aes(x = belong, y = percentage)) + 
  geom_col(color = "azure4", fill = "gainsboro", stat="identity") + 
  geom_text(aes(label = paste0(round(percentage, 1), "%")), 
            position = position_stack(vjust = 0.5), 
            color = "black", size = 4) +
  labs(
    title = "Out of School Event Belong Distribution", 
    subtitle = "Overall", 
    caption = "Source: WiSTEM 2022-2023",
    x = "Belong Responses", 
    y = "Percentage"
    ) +
  theme_minimal() + 
  theme(plot.title = element_text(face = "bold"))

oose_data %>%
  filter(!is.na(belong)) %>%
  count(belong) %>%
  mutate(
    proportion = n / sum(n), 
    percent = (n / sum(n)) * 100 
  ) %>%
  arrange(desc(n)) %>%
  rename(count = n) %>%
  kbl(caption = "Out of School Activity Distribution for belong") %>%
  kable_styling()
```
